BUG: Percent parameters for the mean and standard deviation must be float type.

+++ Examples/Segmentation/VoronoiSegmentation.java
@@ -26,8 +26,8 @@ public class VoronoiSegmentation
     itkVoronoiSegmentationImageFilterUC2UC2UC2_Pointer filter = 
       itkVoronoiSegmentationImageFilterUC2UC2UC2.itkVoronoiSegmentationImageFilterUC2UC2UC2_New();
 
-    filter.SetMeanPercentError( Integer.parseInt( argv[3] ) );
-    filter.SetSTDPercentError(  Integer.parseInt( argv[4] ) );
+    filter.SetMeanPercentError( Double.parseDouble( argv[3] ) );
+    filter.SetSTDPercentError(  Double.parseDouble( argv[4] ) );
 
     itkImageFileWriterUC2_Pointer writer = itkImageFileWriterUC2.itkImageFileWriterUC2_New();
